+++ Collection/Immutable/Enhanced/Map.php
@@ -9,6 +9,13 @@ class Map extends \MF\Collection\Immutable\Map
     /** @var CallbackParser */
     private $callbackParser;
 
+    public static function create(iterable $source, $creator)
+    {
+        $creator = (new CallbackParser())->parseArrowFunction($creator);
+
+        return parent::create($source, $creator);
+    }
+
     public function __construct()
     {
         parent::__construct();