+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There is always Unreleased section on the top. Subsections (Added, Changed, Fixed, Removed) should be added as needed. -->
 ## Unreleased
 - Added `implode` method to `IList`
+- Added `create` method to `ICollection` (_it allows to create collection by callback_)
 
 ## 3.0.1 - 2017-12-20
 - [_dev only_] Added `giorgiosironi/eris` for `Property Based Testing`